+++ packages/migrate/src/migrations/v0.43/zoned-date-time-to-utc-date-time.ts
@@ -10,15 +10,15 @@ import {
 export const migrateZonedDateTimeToUtcDateTime = createContentMigration({
   name: "Migrate zonedDateTime to utcDateTime",
   kind: MigrationKind.AstContentMigration,
-  from: "0.42.0",
-  to: "0.43.0",
+  from: "0.42",
+  to: "0.43",
   migrate: (
     { printNode, printNodes }: MigrationContext,
-    compilerV37: TypeSpecCompilerV0_42,
+    compilerV42: TypeSpecCompilerV0_42,
     root: TypeSpecScriptNode
   ) => {
     const actions: AstContentMigrateAction[] = [];
-    visitRecursive(compilerV37, root, (node) => {
+    visitRecursive(compilerV42, root, (node) => {
       if (
         node.kind === SyntaxKind.TypeReference &&
         node.target.kind === SyntaxKind.Identifier &&