+++ packages/openapi3/test/primitive-types.test.ts
@@ -1,4 +1,5 @@
 import { deepStrictEqual, ok } from "assert";
+import { OpenAPI3Schema } from "../src/types.js";
 import { oapiForModel } from "./test-host.js";
 
 describe("openapi3: primitives", () => {
@@ -24,6 +25,8 @@ describe("openapi3: primitives", () => {
       ["plainTime", { type: "string", format: "time" }],
       ["duration", { type: "string", format: "duration" }],
       ["bytes", { type: "string", format: "byte" }],
+      ["decimal", { type: "string", format: "decimal" }],
+      ["decimal128", { type: "string", format: "decimal128" }],
     ];
 
     for (const [name, expectedSchema] of cases) {
@@ -124,7 +127,7 @@ describe("openapi3: primitives", () => {
   });
 
   describe("using @doc decorator", () => {
-    it("apply description on extended primitive (string)", async () => {
+    it("apply description on extended scalar (string)", async () => {
       const res = await oapiForModel(
         "shortString",
         `
@@ -134,14 +137,13 @@ describe("openapi3: primitives", () => {
       );
 
       ok(res.isRef);
-      ok(res.schemas.shortString, "expected definition named shortString");
       deepStrictEqual(res.schemas.shortString, {
         type: "string",
         description: "My custom description",
       });
     });
 
-    it("apply description on extended primitive (int32)", async () => {
+    it("apply description on extended scalar (int32)", async () => {
       const res = await oapiForModel(
         "specialint",
         `
@@ -151,11 +153,30 @@ describe("openapi3: primitives", () => {
       );
 
       ok(res.isRef);
-      ok(res.schemas.specialint, "expected definition named shortString");
       deepStrictEqual(res.schemas.specialint, {
         type: "integer",
         format: "int32",
         description: "My custom description",
+      });
+    });
+
+    it("apply description on extended custom scalars", async () => {
+      const res = await oapiForModel(
+        "superSpecialint",
+        `
+      @doc("My custom description")
+      scalar specialint extends int32;
+
+      @doc("Override specialint description")
+      scalar superSpecialint extends specialint;
+      `
+      );
+
+      ok(res.isRef);
+      deepStrictEqual(res.schemas.superSpecialint, {
+        type: "integer",
+        format: "int32",
+        description: "Override specialint description",
       });
     });
   });
@@ -188,6 +209,56 @@ describe("openapi3: primitives", () => {
         type: "string",
         format: "password",
       });
+    });
+  });
+
+  describe("using @encode decorator", () => {
+    async function testEncode(
+      scalar: string,
+      expectedOpenApi: OpenAPI3Schema,
+      encoding?: string,
+      encodeAs?: string
+    ) {
+      const encodeAsParam = encodeAs ? `, ${encodeAs}` : "";
+      const encodeDecorator = encoding ? `@encode("${encoding}"${encodeAsParam})` : "";
+      const res = await oapiForModel("s", `${encodeDecorator} scalar s extends ${scalar};`);
+      deepStrictEqual(res.schemas.s, expectedOpenApi);
+    }
+
+    describe("utcDateTime", () => {
+      it("set format to 'date-time' by default", () =>
+        testEncode("utcDateTime", { type: "string", format: "date-time" }));
+      it("set format to 'date-time-rfc7231' when encoding is rfc7231", () =>
+        testEncode("utcDateTime", { type: "string", format: "date-time-rfc7231" }, "rfc7231"));
+
+      it("set type to integer and format to 'unixTimeStamp' when encoding is unixTimestamp", () =>
+        testEncode(
+          "utcDateTime",
+          { type: "integer", format: "unix-timestamp" },
+          "unixTimestamp",
+          "int32"
+        ));
+    });
+
+    describe("offsetDateTime", () => {
+      it("set format to 'date-time' by default", () =>
+        testEncode("offsetDateTime", { type: "string", format: "date-time" }));
+      it("set format to 'date-time-rfc7231' when encoding is rfc7231", () =>
+        testEncode("offsetDateTime", { type: "string", format: "date-time-rfc7231" }, "rfc7231"));
+    });
+
+    describe("duration", () => {
+      it("set format to 'duration' by default", () =>
+        testEncode("duration", { type: "string", format: "duration" }));
+      it("set interger with seconds format setting duration as seconds", () =>
+        testEncode("duration", { type: "integer", format: "seconds" }, "seconds", "int32"));
+    });
+
+    describe("bytes", () => {
+      it("set format to 'base64' by default", () =>
+        testEncode("bytes", { type: "string", format: "byte" }));
+      it("set interger with seconds format setting duration as seconds", () =>
+        testEncode("bytes", { type: "string", format: "base64url" }, "base64url"));
     });
   });
 });