+++ packages/playground/src/components/editor-command-bar.tsx
@@ -1,23 +1,17 @@
 import { Link, Toolbar, ToolbarButton, Tooltip } from "@fluentui/react-components";
 import { Bug16Regular, Save16Regular } from "@fluentui/react-icons";
-import { FunctionComponent, useCallback } from "react";
-import { useSetRecoilState } from "recoil";
-import { SampleConfig } from "../index.js";
-import { PlaygroundManifest } from "../manifest.js";
-import { selectedEmitterState } from "../state.js";
+import { FunctionComponent } from "react";
 import { EmitterDropdown } from "./emitter-dropdown.js";
 import { SamplesDropdown } from "./samples-dropdown.js";
 
 export interface EditorCommandBarProps {
   documentationUrl?: string;
   saveCode: () => Promise<void> | void;
-  updateTypeSpec: (value: string) => void;
   newIssue: () => Promise<void> | void;
 }
 export const EditorCommandBar: FunctionComponent<EditorCommandBarProps> = ({
   documentationUrl,
   saveCode,
-  updateTypeSpec,
   newIssue,
 }) => {
   const documentation = documentationUrl ? (
@@ -27,19 +21,6 @@ export const EditorCommandBar: FunctionComponent<EditorCommandBarProps> = ({
       </Link>
     </label>
   ) : undefined;
-
-  const setEmitter = useSetRecoilState(selectedEmitterState);
-
-  const onSelectSample = useCallback(
-    (config: SampleConfig) => {
-      if (!config.content) throw new Error("Unreachable: sample has no 'content' property");
-
-      updateTypeSpec(config.content);
-
-      setEmitter(config.preferredEmitter ?? PlaygroundManifest.defaultEmitter);
-    },
-    [setEmitter, updateTypeSpec]
-  );
 
   return (
     <div css={{ borderBottom: "1px solid #f5f5f5" }}>
@@ -52,7 +33,7 @@ export const EditorCommandBar: FunctionComponent<EditorCommandBarProps> = ({
             onClick={saveCode as any}
           />
         </Tooltip>
-        <SamplesDropdown onSelectSample={onSelectSample} />
+        <SamplesDropdown />
         <EmitterDropdown />
         {documentation}
         <div css={{ flex: "1" }}></div>