+++ packages/ref-doc/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./api-docs.js";
 export * from "./experimental.js";
 export * from "./extractor.js";
+export * from "./lib.js";