+++ packages/openapi3/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @typespec/openapi3
 
-This log was last generated on Tue, 11 Apr 2023 18:49:17 GMT and should not be manually modified.
+This log was last generated on Wed, 10 May 2023 21:24:00 GMT and should not be manually modified.
+
+## 0.44.0
+Wed, 10 May 2023 21:24:00 GMT
+
+### Updates
+
+- **Added** support for `@encode` decorator
+- Fix openapi3 emitter to mark request body required
+- Support multiple responses for the same status code and content type.
+- Fix issue where shared request bodies did not emit correctly.
+- Update decorator declaration to use `Model` instead of `object`
+- Update dependencies
 
 ## 0.43.0
 Tue, 11 Apr 2023 18:49:17 GMT